#1990 - #2195 merge fixes
