+++ packages/ketcher-core/src/application/editor/actions/template.ts
@@ -16,9 +16,9 @@
 
 import { Atom, Vec2 } from 'domain/entities'
 import { AtomAdd, BondAdd, CalcImplicitH } from '../operations'
-import { atomForNewBond, atomGetAttr, extraBondAction } from './utils'
+import { atomForNewBond, atomGetAttr } from './utils'
 import { fromAtomsAttrs, mergeSgroups } from './atom'
-import { fromBondStereoUpdate, fromBondsAttrs } from './bond'
+import { fromBondStereoUpdate, fromBondsAttrs, fromBondAddition } from './bond'
 
 import { Action } from './action'
 import closest from '../shared/closest'
@@ -38,6 +38,43 @@ export function fromTemplateOnCanvas(restruct, template, pos, angle) {
   action.addOp(new CalcImplicitH(pasteItems.atoms).perform(restruct))
 
   return [action, pasteItems]
+}
+
+function extraBondAction(restruct, aid, angle) {
+  let action = new Action()
+  const frid = atomGetAttr(restruct, aid, 'fragment')
+  let additionalAtom: any = null
+
+  if (angle === null) {
+    const middleAtom = atomForNewBond(restruct, aid)
+    const actionRes = fromBondAddition(
+      restruct,
+      { type: 1 },
+      aid,
+      middleAtom.atom,
+      middleAtom.pos.get_xy0()
+    )
+    action = actionRes[0]
+    action.operations.reverse()
+    additionalAtom = actionRes[2]
+  } else {
+    const operation = new AtomAdd(
+      { label: 'C', fragment: frid },
+      new Vec2(1, 0)
+        .rotate(angle)
+        .add(restruct.molecule.atoms.get(aid).pp)
+        .get_xy0()
+    ).perform(restruct) as AtomAdd
+
+    action.addOp(operation)
+    action.addOp(
+      new BondAdd(aid, operation.data.aid, { type: 1 }).perform(restruct)
+    )
+
+    additionalAtom = operation.data.aid
+  }
+
+  return { action, aid1: additionalAtom }
 }
 
 export function fromTemplateOnAtom(restruct, template, aid, angle, extraBond) {
